+++ services/src/api/doc.rs
@@ -1,3 +1,13 @@
+use super::handlers;
+use super::handlers::plots::WrappedPlotOutput;
+use super::handlers::spatial_references::{AxisLabels, AxisOrder, SpatialReferenceSpecification};
+use super::handlers::tasks::{TaskAbortOptions, TaskResponse};
+use super::handlers::upload::{UploadFileLayersResponse, UploadFilesResponse};
+use super::handlers::wcs::CoverageResponse;
+use super::handlers::wfs::{CollectionType, Coordinates, Feature, FeatureType, GeoJson};
+use super::handlers::wms::MapResponse;
+use super::handlers::workflows::RasterStreamWebsocketResultType;
+use super::model::responses::ErrorResponse;
 use crate::api::model::datatypes::{
     BoundingBox2D, Breakpoint, ClassificationMeasurement, Colorizer, ContinuousMeasurement,
     Coordinate2D, DataId, DataProviderId, DatasetId, DatasetName, DateTimeParseFormat,
@@ -26,28 +36,17 @@ use crate::api::model::responses::{
 use crate::api::model::services::{
     AddDataset, CreateDataset, DataPath, DatasetDefinition, MetaDataDefinition, MetaDataSuggestion,
 };
+use crate::api::ogc::{util::OgcBoundingBox, wcs, wfs, wms};
 use crate::contexts::{SessionId, SimpleSession};
 use crate::datasets::listing::{DatasetListing, OrderBy, Provenance, ProvenanceOutput};
 use crate::datasets::storage::{AutoCreateDataset, Dataset};
 use crate::datasets::upload::{UploadId, Volume, VolumeName};
 use crate::datasets::{RasterDatasetFromWorkflow, RasterDatasetFromWorkflowResult};
-use crate::handlers;
-use crate::handlers::plots::WrappedPlotOutput;
-use crate::handlers::spatial_references::{AxisLabels, AxisOrder, SpatialReferenceSpecification};
-use crate::handlers::tasks::{TaskAbortOptions, TaskResponse};
-use crate::handlers::upload::{UploadFileLayersResponse, UploadFilesResponse};
-use crate::handlers::wcs::CoverageResponse;
-use crate::handlers::wfs::{CollectionType, Coordinates, Feature, FeatureType, GeoJson};
-use crate::handlers::wms::MapResponse;
-use crate::handlers::workflows::RasterStreamWebsocketResultType;
-use crate::handlers::ErrorResponse;
 use crate::layers::layer::{
     AddLayer, AddLayerCollection, CollectionItem, Layer, LayerCollection, LayerCollectionListing,
     LayerListing, Property, ProviderLayerCollectionId, ProviderLayerId,
 };
 use crate::layers::listing::LayerCollectionId;
-use crate::ogc::util::OgcBoundingBox;
-use crate::ogc::{wcs, wfs, wms};
 use crate::projects::{
     ColorParam, CreateProject, DerivedColor, DerivedNumber, LayerUpdate, LayerVisibility,
     LineSymbology, NumberParam, Plot, PlotUpdate, PointSymbology, PolygonSymbology, Project,
