+++ services/src/api/handlers/plots.rs
@@ -1,9 +1,8 @@
 use crate::api::model::datatypes::TimeInterval;
-use crate::contexts::ApplicationContext;
+use crate::api::ogc::util::{parse_bbox, parse_time};
+use crate::contexts::{ApplicationContext, SessionContext};
 use crate::error;
 use crate::error::Result;
-use crate::handlers::SessionContext;
-use crate::ogc::util::{parse_bbox, parse_time};
 use crate::util::config;
 use crate::util::parsing::parse_spatial_resolution;
 use crate::util::server::connection_closed;
