+++ services/src/api/handlers/spatial_references.rs
