+++ services/src/api/handlers/wms.rs
@@ -5,25 +5,26 @@ use reqwest::Url;
 use serde_json::json;
 use snafu::{ensure, ResultExt};
 
+use crate::api::model::datatypes::{SpatialReference, SpatialReferenceOption, TimeInterval};
+use crate::api::ogc::util::{ogc_endpoint_url, OgcProtocol, OgcRequestGuard};
+use crate::api::ogc::wms::request::{
+    GetCapabilities, GetLegendGraphic, GetMap, GetMapExceptionFormat,
+};
+use crate::contexts::ApplicationContext;
+use crate::contexts::SessionContext;
+use crate::error::Result;
+use crate::error::{self, Error};
+use crate::util::config;
+use crate::util::config::get_config_element;
+use crate::util::server::{connection_closed, not_implemented_handler, CacheControlHeader};
+use crate::workflows::registry::WorkflowRegistry;
+use crate::workflows::workflow::WorkflowId;
 use geoengine_datatypes::primitives::{
     AxisAlignedRectangle, RasterQueryRectangle, SpatialPartition2D,
 };
 use geoengine_datatypes::{operations::image::Colorizer, primitives::SpatialResolution};
 use utoipa::openapi::{KnownFormat, ObjectBuilder, SchemaFormat, SchemaType};
 use utoipa::ToSchema;
-
-use crate::api::model::datatypes::{SpatialReference, SpatialReferenceOption, TimeInterval};
-use crate::contexts::ApplicationContext;
-use crate::error::Result;
-use crate::error::{self, Error};
-use crate::handlers::SessionContext;
-use crate::ogc::util::{ogc_endpoint_url, OgcProtocol, OgcRequestGuard};
-use crate::ogc::wms::request::{GetCapabilities, GetLegendGraphic, GetMap, GetMapExceptionFormat};
-use crate::util::config;
-use crate::util::config::get_config_element;
-use crate::util::server::{connection_closed, not_implemented_handler, CacheControlHeader};
-use crate::workflows::registry::WorkflowRegistry;
-use crate::workflows::workflow::WorkflowId;
 
 use geoengine_operators::engine::{
     CanonicOperatorName, ExecutionContext, ResultDescriptor, SingleRasterOrVectorSource,
@@ -477,8 +478,8 @@ fn default_time_from_config() -> TimeInterval {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::api::model::responses::ErrorResponse;
     use crate::contexts::{Session, SimpleApplicationContext};
-    use crate::handlers::ErrorResponse;
     use crate::util::tests::with_temp_context;
     use crate::util::tests::{
         check_allowed_http_methods, read_body_string, register_ndvi_workflow_helper,
@@ -624,7 +625,7 @@ mod tests {
             // geoengine_datatypes::util::test::save_test_bytes(&image_bytes, "raster_small.png");
 
             assert_eq!(
-                include_bytes!("../../../test_data/wms/raster_small.png") as &[u8],
+                include_bytes!("../../../../test_data/wms/raster_small.png") as &[u8],
                 image_bytes.as_slice()
             );
         })
@@ -667,7 +668,7 @@ mod tests {
         // geoengine_datatypes::util::test::save_test_bytes(&image_bytes, "get_map.png");
 
         assert_eq!(
-            include_bytes!("../../../test_data/wms/get_map.png") as &[u8],
+            include_bytes!("../../../../test_data/wms/get_map.png") as &[u8],
             image_bytes
         );
     }
@@ -695,7 +696,7 @@ mod tests {
         // geoengine_datatypes::util::test::save_test_bytes(&image_bytes, "get_map_ndvi.png");
 
         assert_eq!(
-            include_bytes!("../../../test_data/wms/get_map_ndvi.png") as &[u8],
+            include_bytes!("../../../../test_data/wms/get_map_ndvi.png") as &[u8],
             image_bytes
         );
 
@@ -733,7 +734,7 @@ mod tests {
         // geoengine_datatypes::util::test::save_test_bytes(&image_bytes, "get_map.png");
 
         assert_eq!(
-            include_bytes!("../../../test_data/wms/get_map.png") as &[u8],
+            include_bytes!("../../../../test_data/wms/get_map.png") as &[u8],
             image_bytes
         );}).await;
     }
@@ -821,7 +822,7 @@ mod tests {
                 // geoengine_datatypes::util::test::save_test_bytes(&image_bytes, "get_map_colorizer.png");
 
                 assert_eq!(
-                    include_bytes!("../../../test_data/wms/get_map_colorizer.png") as &[u8],
+                    include_bytes!("../../../../test_data/wms/get_map_colorizer.png") as &[u8],
                     image_bytes
                 );
             },
