+++ services/src/error.rs
@@ -1,9 +1,9 @@
 use crate::api::model::datatypes::{
     DataProviderId, DatasetId, LayerId, SpatialReference, SpatialReferenceOption, TimeInstance,
 };
+use crate::api::model::responses::ErrorResponse;
 use crate::datasets::external::aruna::error::ArunaProviderError;
 use crate::datasets::external::netcdfcf::NetCdfCf4DProviderError;
-use crate::handlers::ErrorResponse;
 use crate::{layers::listing::LayerCollectionId, workflows::workflow::WorkflowId};
 use actix_web::http::StatusCode;
 use actix_web::HttpResponse;
@@ -354,7 +354,7 @@ pub enum Error {
 
     #[snafu(context(false))]
     WorkflowApi {
-        source: crate::handlers::workflows::WorkflowApiError,
+        source: crate::api::handlers::workflows::WorkflowApiError,
     },
 
     SubPathMustNotEscapeBasePath {
