+++ services/src/pro/apidoc.rs
@@ -1,3 +1,14 @@
+use crate::api::handlers;
+use crate::api::handlers::plots::WrappedPlotOutput;
+use crate::api::handlers::spatial_references::{
+    AxisLabels, AxisOrder, SpatialReferenceSpecification,
+};
+use crate::api::handlers::tasks::{TaskAbortOptions, TaskResponse};
+use crate::api::handlers::upload::{UploadFileLayersResponse, UploadFilesResponse};
+use crate::api::handlers::wcs::CoverageResponse;
+use crate::api::handlers::wfs::{CollectionType, Coordinates, Feature, FeatureType, GeoJson};
+use crate::api::handlers::wms::MapResponse;
+use crate::api::handlers::workflows::RasterStreamWebsocketResultType;
 use crate::api::model::datatypes::{
     BoundingBox2D, Breakpoint, ClassificationMeasurement, Colorizer, ContinuousMeasurement,
     Coordinate2D, DataId, DataProviderId, DatasetId, DatasetName, DateTime, DateTimeParseFormat,
@@ -19,6 +30,7 @@ use crate::api::model::operators::{
     VectorResultDescriptor,
 };
 use crate::api::model::responses::datasets::DatasetNameResponse;
+use crate::api::model::responses::ErrorResponse;
 use crate::api::model::responses::{
     BadRequestQueryResponse, IdResponse, PayloadTooLargeResponse, UnauthorizedAdminResponse,
     UnauthorizedUserResponse, UnsupportedMediaTypeForJsonResponse,
@@ -26,28 +38,17 @@ use crate::api::model::responses::{
 use crate::api::model::services::{
     AddDataset, CreateDataset, DataPath, DatasetDefinition, MetaDataDefinition, MetaDataSuggestion,
 };
+use crate::api::ogc::{util::OgcBoundingBox, wcs, wfs, wms};
 use crate::contexts::SessionId;
 use crate::datasets::listing::{DatasetListing, OrderBy, Provenance, ProvenanceOutput};
 use crate::datasets::storage::{AutoCreateDataset, Dataset};
 use crate::datasets::upload::{UploadId, Volume, VolumeName};
 use crate::datasets::{RasterDatasetFromWorkflow, RasterDatasetFromWorkflowResult};
-use crate::handlers;
-use crate::handlers::plots::WrappedPlotOutput;
-use crate::handlers::spatial_references::{AxisLabels, AxisOrder, SpatialReferenceSpecification};
-use crate::handlers::tasks::{TaskAbortOptions, TaskResponse};
-use crate::handlers::upload::{UploadFileLayersResponse, UploadFilesResponse};
-use crate::handlers::wcs::CoverageResponse;
-use crate::handlers::wfs::{CollectionType, Coordinates, Feature, FeatureType, GeoJson};
-use crate::handlers::wms::MapResponse;
-use crate::handlers::workflows::RasterStreamWebsocketResultType;
-use crate::handlers::ErrorResponse;
 use crate::layers::layer::{
     AddLayer, AddLayerCollection, CollectionItem, Layer, LayerCollection, LayerCollectionListing,
     LayerListing, Property, ProviderLayerCollectionId, ProviderLayerId,
 };
 use crate::layers::listing::LayerCollectionId;
-use crate::ogc::util::OgcBoundingBox;
-use crate::ogc::{wcs, wfs, wms};
 use crate::pro;
 use crate::pro::handlers::users::{Quota, UpdateQuota};
 use crate::projects::{
