+++ services/src/pro/tasks/mod.rs
@@ -14,9 +14,9 @@ use super::users::UserSession;
 
 // TODO: implement real permissions on task types
 const ADMIN_ONLY_TASKS: [&str; 3] = [
-    crate::handlers::ebv::EBV_OVERVIEW_TASK_TYPE,
-    crate::handlers::ebv::EBV_MULTI_OVERVIEW_TASK_TYPE,
-    crate::handlers::ebv::EBV_REMOVE_OVERVIEW_TASK_TYPE,
+    crate::api::handlers::ebv::EBV_OVERVIEW_TASK_TYPE,
+    crate::api::handlers::ebv::EBV_MULTI_OVERVIEW_TASK_TYPE,
+    crate::api::handlers::ebv::EBV_REMOVE_OVERVIEW_TASK_TYPE,
 ];
 
 #[derive(Default)]
